ai_client_app/requirements.txt
PyQt6
markdown
openai
anthropic
python-dotenv